data cleaning uninsured population
I have cleaned out most of the unnecessary metadata but the household income and the ratio of the poverty level as I am still hesitating to delete the latter one or that we should keep it. Besides that I have cleaned out the margin of errors and kept the total population in the columns. Besides that, I have stored my environment with the (almost) cleaned datasets of the uninsured population in the US of the years 2010-2023 except for year 2020