class coercion
The values in the datasets with the uninsured population numbers of the states per year were stored as strings/characters. As we want to compute the share of population unsinsured we have to be able to work with numeric. That is why I have coerced these strings into numeric values by deleting the "" and % characters and then I could coerce them into numeric values.