making new variable share uninsured population
Here I have made a new variable, named share uninsured population. After every column of the state with the absolute values of the uninsured population there is a new column with the share of the uninsured population in a state. I have done this for every year. this share is calculated by dividing the uninsured population of a given state by the total civilian noninstitutionalized population in a given year. Besides that, I have stored the environment with the datasets including these new variables and are called for example df_pop_2010